export fixings

+++ vedo/picture.py
@@ -845,7 +845,7 @@ class Picture(vtk.vtkImageActor, vedo.base.Base3DProp):
         nx, ny, _ = self._data.GetDimensions()
         nchan = self._data.GetPointData().GetScalars().GetNumberOfComponents()
         narray = utils.vtk2numpy(self._data.GetPointData().GetScalars()).reshape(ny,nx,nchan)
-        narray = np.flip(narray, axis=0)
+        narray = np.flip(narray, axis=0).astype(np.uint8)
         return narray
 
     def box(self, xspan, yspan, c='green5', alpha=1):
